added product crud

+++ back/api/src/database/migrations/2024_03_12_164015_create_product_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->decimal('price');
+            $table->double('price');
             $table->unsignedBigInteger('category_id')->nullable();
             $table->unsignedBigInteger('vendor_id')->nullable();
             $table->unsignedBigInteger('brand_id')->nullable();
